Commit project
Currently the project is functional. There is a delay when selecting a new base color, where the clock face background doesn't update until the next set interval run (at least I think this is the case).

In the original Vue project this was instantaneous so I need to figure out what is different about React's render triggers